requirements.txt
six
nose
sphinx
flask
connexion
flask_injector
mysqlclient
flask-sqlalchemy
